+++ src/postage/mod.rs
@@ -19,7 +19,10 @@ pub use stamp_math::{
     EFFECTIVE_SIZE_BREAKPOINTS, get_depth_for_size, get_stamp_cost, get_stamp_effective_bytes,
     get_stamp_theoretical_bytes, get_stamp_usage,
 };
-pub use stamper::{Envelope, MIN_DEPTH, NUM_BUCKETS, Stamper};
+pub use stamper::{
+    Envelope, MARSHALED_STAMP_LENGTH, MIN_DEPTH, NUM_BUCKETS, Stamper,
+    convert_envelope_to_marshaled_stamp, marshal_stamp,
+};
 pub use types::{BatchBucket, GlobalPostageBatch, PostageBatch, PostageBatchBuckets};
 
 /// Handle exposing the postage endpoints. Cheap to clone.